fix: corrective pass for B1/B2/N1/N2/N4 from remaining-scope review
- B1: Restored 25 stripped # noqa: F401/F841 directives (shared/pptx/__init__.py,
  shared/pptx/pattern_injectors/__init__.py, scripts/package_audit.py,
  tests/test_pattern_assets.py). Fixed I001 import ordering in
  shared/pptx/layouts.py and shared/pptx/pattern_injectors/registry.py
  (from collections.abc before from typing). ruff check . now green
  (only pre-existing analyzer.py I001 in untracked dir).
- B2: Changed mermaid-smoke CI install to 'pip install -e ".[dev]"' so
  pytest is available.
- N1: Added real _CONTENT_MARKERS and _MIN_CONTENT_SHAPES for the 5 new
  families (kpi-dashboard-grid: 23 shapes, maturity-model-ladder: 17,
  case-study-card: 16, checklist-status: 20, quote-testimonial-card: 12)
  with markers verified from known-good builds.
- N2: Updated stale docstrings: module header (6 families -> 11), class
  docstring (removed 'not built' note), assertion message (Expected 6 -> 11).
- N4: Restored LF line endings in tests/test_media_library.py and
  tests/test_taxonomy_sync.py (were CRLF). Clean 1-line diffs each.
- Verified: ruff check ., pytest -q (585 passed, 0 failed, 6 xfailed),
  release_gate.py (all 12 steps PASSED).

+++ shared/pptx/style.py
@@ -74,4 +74,4 @@ def no_line(shape):
 
 def inches(value: Any) -> int:
     """EMU from inches (accepts int/float). 914400 EMU = 1 inch."""
-    return int(round(float(value) * 914400))
+    return round(float(value) * 914400)
